Moved `SourceBuffer` to `joe.scanner`.
