+++ lib/src/main/java/com/wjduquette/joe/Trace.java
@@ -1,6 +1,6 @@
 package com.wjduquette.joe;
 
-import com.wjduquette.joe.SourceBuffer.Span;
+import com.wjduquette.joe.scanner.SourceBuffer.Span;
 
 /**
  * A message associated with a JoeError, possibly with a script location.
