+++ lib/src/main/java/com/wjduquette/joe/walker/Interpreter.java
@@ -2,6 +2,7 @@ package com.wjduquette.joe.walker;
 
 import com.wjduquette.joe.*;
 import com.wjduquette.joe.patterns.Matcher;
+import com.wjduquette.joe.scanner.SourceBuffer;
 import com.wjduquette.joe.types.ListValue;
 import com.wjduquette.joe.types.MapValue;
