+++ lib/src/main/java/com/wjduquette/joe/walker/Scanner.java
@@ -1,6 +1,7 @@
 package com.wjduquette.joe.walker;
 
 import com.wjduquette.joe.*;
+import com.wjduquette.joe.scanner.SourceBuffer;
 
 import java.util.*;
