+++ lib/src/test/java/com/wjduquette/joe/walker/ScannerTest.java
@@ -1,6 +1,6 @@
 package com.wjduquette.joe.walker;
 
-import com.wjduquette.joe.SourceBuffer;
+import com.wjduquette.joe.scanner.SourceBuffer;
 import com.wjduquette.joe.Ted;
 import org.junit.Test;
